requirements.txt
fastapi==0.115.12
uvicorn==0.27.1
python-multipart==0.0.9
pillow==10.2.0
tensorflow==2.18.0
numpy==2.0.2
requests==2.31.0
python-dotenv==1.0.1
kagglehub
streamlit==1.44.0